Improve alerts deduplication filtering
- Problem with Flink: left joins sometimes give early results wich are retracted few milliseconds later. This creates
  short term alerts on Alerta which creates sometimes a lot of unneccessary alerts flood and confuses the user.
- This is solved by adding Flink later arrival filtering: Alerts which are 200ms late are grouped with the previous window.
- This creates on the other hand a more complicated E2E testing case.

+++ test/bats/test-core-sql-services/test-alerta-bulk-filter.bats
@@ -3,45 +3,46 @@
 if [ -z "${SELF_HOSTED_RUNNER}" ]; then
     SUDO="sudo -E"
 fi
-
+DEBUG=${DEBUG:-false} # set this to true to disable starting and stopping of kubefwd
 ALERT1=/tmp/ALERT1
 ALERT2=/tmp/ALERT2
 ALERT3=/tmp/ALERT3
 ALERT4=/tmp/ALERT4
+ALERT5=/tmp/ALERT5
+ALERT6=/tmp/ALERT6
 ALERTS_TOPIC=iff.alerts
 BULK_ALERTS_TOPIC=iff.alerts.bulk
 KAFKACAT_ATTRIBUTES=/tmp/KAFKACAT_ATTRIBUTES
 KAFKACAT_ATTRIBUTES_FILTERED=/tmp/KAFKACAT_ATTRIBUTES_FILTERED
 KAFKA_BOOTSTRAP=my-cluster-kafka-bootstrap:9092
 
+
 cat << EOF | tr -d '\n' > ${ALERT1}
-{
+{"resource":"urn:plasmacutter-test:12345","event":"Test event"};{
   "resource": "urn:plasmacutter-test:12345",
   "event": "Test event",
   "environment": "Development",
-  "service": [
-    "E2E test"
-  ],
+  "service": ["E2E test"],
   "severity": "ok",
   "customer": "test",
   "text": "OK"
 }
 EOF
+
 cat << EOF | tr -d '\n' > ${ALERT2}
-{
+{"resource":"urn:plasmacutter-test:12345","event":"Test event"};{
   "resource": "urn:plasmacutter-test:12345",
   "event": "Test event",
   "environment": "Development",
-  "service": [
-    "E2E test"
-  ],
+  "service": ["E2E test"],
   "severity": "warning",
   "customer": "test",
   "text": "Warning"
 }
 EOF
+
 cat << EOF | tr -d '\n' > ${ALERT3}
-{
+{"resource":"urn:plasmacutter-test:12345","event":"Test event"};{
   "resource": "urn:plasmacutter-test:12345",
   "event": "Test event",
   "environment": "Development",
@@ -53,8 +54,9 @@ cat << EOF | tr -d '\n' > ${ALERT3}
   "text": "critical"
 }
 EOF
+
 cat << EOF | tr -d '\n' > ${ALERT4}
-{
+{"resource":"urn:plasmacutter-test:12345","event":"Test event"};{
   "resource": "urn:plasmacutter-test:12345",
   "event": "Test event",
   "environment": "Development",
@@ -66,6 +68,35 @@ cat << EOF | tr -d '\n' > ${ALERT4}
   "text": "critical"
 }
 EOF
+
+cat << EOF | tr -d '\n' > ${ALERT5}
+{"resource":"urn:plasmacutter-test:xxxx","event":"Test event"};{
+  "resource": "urn:plasmacutter-test:xxxx",
+  "event": "Test event",
+  "environment": "Development",
+  "service": [
+    "E2E test"
+  ],
+  "severity": "flush",
+  "customer": "test",
+  "text": "critical"
+}
+EOF
+
+cat << EOF | tr -d '\n' > ${ALERT6}
+{"resource":"urn:plasmacutter-test:xxxx","event":"Test event"};{
+  "resource": "urn:plasmacutter-test:xxxx",
+  "event": "Test event",
+  "environment": "Development",
+  "service": [
+    "E2E test"
+  ],
+  "severity": "flush",
+  "customer": "test",
+  "text": "criticalx"
+}
+EOF
+
 compare_attributes1() {
     cat << EOF | diff "$1" - >&3
 {"resource":"urn:plasmacutter-test:12345","event":"Test event","environment":"Development","service":["E2E test"],"severity":"ok","customer":"test","text":"OK"}
@@ -78,30 +109,60 @@ EOF
 
 setup() {
     # shellcheck disable=SC2086
-    (exec ${SUDO} kubefwd -n iff -l app.kubernetes.io/name=kafka svc  >/dev/null 2>&1) &
+    [ $DEBUG = "true" ] || (exec ${SUDO} kubefwd -n iff -l app.kubernetes.io/name=kafka svc  >/dev/null 2>&1) &
     echo "# launched kubefwd for kafka, wait some seconds to give kubefwd to launch the services"
     sleep 2
 }
 teardown(){
     echo "# now killing kubefwd"
     # shellcheck disable=SC2086
-    ${SUDO} killall kubefwd
+    [ $DEBUG = "true" ]  || ${SUDO} killall kubefwd
 }
 
 
+# This tests check whether Flink filters the alert correcty. The filtering
+# uses late data arrival, e.g. every event is allowed to arrive 200ms late
+# and still will be assigned to the window before. This is to make sure that
+# no short term event shows up in the Alerta dashboard.
+# This test looks overcomplicated. Reason is a bug in Flink late data handling
+# It seems not to update the watermark when data only with the same key is sent
+# That is there are two different keys: ALERTS1-4 and ALERTS5-6
 @test "verify only changes of alert sequence from alerts.bulk are forwarded to alerts topic" {
-    (exec stdbuf -oL kafkacat -C -t ${ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -o end >${KAFKACAT_ATTRIBUTES}) &
+    (exec stdbuf -oL kafkacat -C -t ${ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -o end -f '%s\n'>${KAFKACAT_ATTRIBUTES}) &
     sleep 2 # wait for next aggregation window
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT4}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT1}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT1}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT2}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT2}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT1}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT1}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT3}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT3}
-    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} <${ALERT1}
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT4}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT5}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT1}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT6}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT2}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT5}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT1}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT6}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT1}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT5}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT3}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT6}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT3}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT5}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT1}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT6}
+    sleep 0.2
+    kafkacat -P -t ${BULK_ALERTS_TOPIC} -b ${KAFKA_BOOTSTRAP} -K';' <${ALERT1}
     echo "# Sent attribute to attribute topic, wait some time for aggregation"
     sleep 2
     killall kafkacat
